refactor(symphony): centralize session summary format helpers
